Add files via upload
Please find attached 8 codes that correspond to the addition and deletion operation for each data structure. Operations are executed at start,middle and end. Please also find the .xlsx file with the time measures for each code. Please take in account that I focused on getting the times complexity measures that the code build itself due the short time to complete this task and the extension of the codes.